Update the ProductServiceImpl.java file to include a new method for updating product details.

+++ src/main/java/com/ecomproject/request/CreateProductRequest.java
@@ -9,7 +9,7 @@ public class CreateProductRequest {
 
     private String title;
     private String description;
-    private double mrpPrice;
+    private int mrpPrice;
     private int sellingPrice;
     private String color;
     private List<String> images;
